[#141] rename: constant 폴더에 enum 추가
