+++ rest/src/main/java/com/wherecar/rest/announcement/domain/Announcement.java
@@ -1,5 +1,6 @@
 package com.wherecar.rest.announcement.domain;
 
+import com.wherecar.rest.announcement.domain.constant.AnnouncementType;
 import com.wherecar.rest.common.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
