--- rest/src/main/java/com/wherecar/rest/announcement/domain/AnnouncementType.java
@@ -1,6 +1,0 @@
-package com.wherecar.rest.announcement.domain;
-
-public enum AnnouncementType {
-    INFO,
-    ALERT
-}
