+++ rest/src/main/java/com/wherecar/rest/announcement/domain/constant/AnnouncementType.java
@@ -1,0 +1,6 @@
+package com.wherecar.rest.announcement.domain.constant;
+
+public enum AnnouncementType {
+    INFO,
+    ALERT
+}
