+++ rest/src/main/java/com/wherecar/rest/car/domain/constant/CarState.java
@@ -1,4 +1,4 @@
-package com.wherecar.rest.car.domain;
+package com.wherecar.rest.car.domain.constant;
 
 public enum CarState {
     RUNNING,
