+++ rest/src/main/java/com/wherecar/rest/car/domain/constant/OwnerType.java
@@ -1,4 +1,4 @@
-package com.wherecar.rest.car.domain;
+package com.wherecar.rest.car.domain.constant;
 
 public enum OwnerType {
     PERSONAL,
