+++ rest/src/main/java/com/wherecar/rest/carlog/application/dto/CarLogsResponse.java
@@ -1,7 +1,7 @@
 package com.wherecar.rest.carlog.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.wherecar.rest.carlog.domain.DriveType;
+import com.wherecar.rest.carlog.domain.constant.DriveType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
