--- rest/src/main/java/com/wherecar/rest/carlog/domain/DriveType.java
@@ -1,6 +1,0 @@
-package com.wherecar.rest.carlog.domain;
-
-public enum DriveType {
-    COMMUTE,
-    WORK
-}
