+++ rest/src/main/java/com/wherecar/rest/carlog/domain/constant/DriveType.java
@@ -1,0 +1,6 @@
+package com.wherecar.rest.carlog.domain.constant;
+
+public enum DriveType {
+    COMMUTE,
+    WORK
+}
