+++ rest/src/main/java/com/wherecar/rest/user/domain/Permission.java
@@ -1,5 +1,6 @@
 package com.wherecar.rest.user.domain;
 
+import com.wherecar.rest.user.domain.constant.PermissionType;
 import jakarta.persistence.*;
 import lombok.*;
