+++ rest/src/main/java/com/wherecar/rest/user/infrastructure/PermissionRepository.java
@@ -1,7 +1,7 @@
 package com.wherecar.rest.user.infrastructure;
 
 import com.wherecar.rest.user.domain.Permission;
-import com.wherecar.rest.user.domain.PermissionType;
+import com.wherecar.rest.user.domain.constant.PermissionType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
